backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose[cryptography]
python-multipart
celery
redis
